(gemini) download blobs during repo sync (command only)
Prompts:

For the `at-mirror repo sync` command, we also need to fetch the accounts blob files. This will involve adding a step to the process where we fetch the CAR. The blobs should be saved in `data/repos/<did>/blobs/<cid>.blob`

To fetch the list of blobs, use: https://docs.bsky.app/docs/api/com-atproto-sync-list-blobs

To fetch a single blob, use: https://docs.bsky.app/docs/api/com-atproto-sync-get-blob

...

Blobs don't use since, but are still ordered on cursor in the list function. When sync'n blobs, we probably fetch the full list of cid's on first sync, and then check to see if we have the file before downloading it